+++ contracts/community/interfaces/IHenkaku1155Mint.sol
