--- scripts/public/deployed_contract_addr_mumbai.json
@@ -1,4 +1,0 @@
-{
-  "Fowarder": "0x5FbDB2315678afecb367f032d93F642f64180aa3",
-  "Nengajo": "0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512"
-}
